back @testing-library/dom to v8
`@testing-library/dom` is also a dependency of `jest-config-stripes` and
we want it to act like a peer-dep, i.e. we only want a single version in
our bundle so that jest doesn't get confused. It may be that there's a
better way to do configure this, but since we're relying on this
package's deps being hoisted by the UI app that depends on this one,
this seems like the easiest way to accomplish that.