correctly group mod-descriptor conditionals
The module-descriptor should be published if PUBLISH_MOD_DESCRIPTOR is
true and the destination branch is the default-branch (master or main).
IOW, we're evaluating `(A && B) || C` when we want `A && (B || C)`.
Whoops.

This error appears to be widespread but hasn't been picked up because
most UI repos have their default branch as `master`, meaning we end
upwith a conditional like `false && true || false` (which evaluates to
false) but here we end up with `false && true || true (which evaluates
to true).