React v18 compatibility (#7)
* BREAKING @testing-library/dom 9
* BREAKING @testing-library/react 14
* BREAKING @testing-library/user-event 14
* BREAKING jest-junit 16
* BREAKING jsdom 22

Even though we never formally released v1, the initial work here was all
predicated on react v17 versions of things. In order to preserve test
functionality for repositories that haven't yet upgraded to react v18,
we bump the major version here to separate the react v18 changes with a
breaking semver change.