Update yarn.lock to pacify snyk, bump stripes-webpack to 6.0.1. Refs STRIPES-963, STRWEB-130.
Current yarn.lock contains `@folio/stripes-webpack@6.0.10990000000052` › `crypto-browserify@3.12.1` › `pbkdf2@3.1.2`.

pbkdf2 < 3.1.3 has this critical severity security vulnerability:

* CVE-2025-6545 = https://github.com/advisories/GHSA-h7cp-r72f-jxh6

crypto-browserify has been removed from stripes-webpack:

https://folio-org.atlassian.net/browse/STRIPES-963
https://folio-org.atlassian.net/browse/STRWEB-130

Updating yarn.lock removes crypto-browserify and pbkdf2 from yarn.lock.

Bumping stripes-webpack from 6.0.0 to 6.0.1 ensures that 6.0.10990000000052 is not used.

Of course, none of these versions will actually be installed since this repo is only installed as a dependency, but we can still be nice to snyk and help it calm down, and use versions in yarn.lock for local tests that are not outdated but near to the actual used versions.